Refactoring: Divide common logic into Fluent::Logplex
